Finished lab 5. TBB and MP implementations are both faster than the pass_init implementation, but the cilk implementation is not as fast. (By speed I mean faster for long time computations!) This could be partly because I used cilk_spawn rather than cilk_for, but cilk_for was doing some really bizzare things.
Other notes are that I like tbb mostly, and that mp is really really
easy to implement (took just a minute or so). Also, my implementations
are not technically maps, since I have a clear side effect where the
a loop that finds the final hash will write to the variable done.
This doesn't create any race condition (both because done only makes
it so that the loop stops, and that only one process will actually
write to done unless there is an md5 hash collission on 8 digit pins,
which is probably less likely than being killed by a meteor tomorrow.